+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/instrumentation/androidx/room/visitor/Common.kt
@@ -1,3 +1,0 @@
-package io.sentry.android.gradle.instrumentation.androidx.room.visitor
-
-const val DESCRIPTION = "room query with mapping"