+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix: Correctly add the proguard UUID output directory to the source set (#226)
 * Feature: Make the ignoreXXX properties in SentryPluginExtension sets (#225)
 * Expose SentryPluginExtension.tracingInstrumentation (#229)
+* Ref: Change Room queries description to Dao class name (#232)
 
 ## 3.0.0-beta.1